refacoring of UI
